Move TeXiFy pdf viewer implementation to this plugin
