+++ src/main/resources/META-INF/texify-idea.xml
@@ -1,0 +1,5 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="nl.rubensten.texifyidea">
+        <pdfViewer implementation="com.firsttimeinforever.intellij.pdf.viewer.tex.TexPdfViewer"/>
+    </extensions>
+</idea-plugin>
